Logo dimensions fix
Fixed the weird logo dimensions bug which is present on
http://www.jsisrael.com.

It occurred because the Pivot template's JS code tried to detect the
logo's aspect ratio before the image was loaded.

We can't detect the logo dimensions until the actual image is loaded,
which happens after `$(document).ready()` since it's a remote resource.

To work around this, we can hard-code the `logo-wide` class, since we
know that our logo is wide (and not square) in its dimensions.